+++ pytest_rerunfailures.py
@@ -4,7 +4,9 @@ import warnings
 
 import pkg_resources
 import pytest
+from _pytest.reports import TestReport
 from _pytest.runner import runtestprotocol
+from _pytest.runner import SetupState
 
 HAS_RESULTLOG = False
 
@@ -20,6 +22,13 @@ except ImportError:
 PYTEST_GTE_54 = pkg_resources.parse_version(
     pytest.__version__
 ) >= pkg_resources.parse_version("5.4")
+
+
+class RerunAwareSetupState(SetupState):
+    def teardown_exact(self, item, nextitem):
+        if item.will_be_rerun:
+            nextitem = item
+        return super().teardown_exact(item, nextitem)
 
 
 def works_with_current_xdist():
@@ -79,6 +88,18 @@ def pytest_configure(config):
         "to 'reruns' times. Add a delay of 'reruns_delay' seconds "
         "between re-runs.",
     )
+
+
+def pytest_collection_modifyitems(items):
+    for item in items:
+        item.execution_count = 0
+        item.allowed_reruns = get_reruns_count(item) or 0
+        item.will_be_rerun = False
+
+
+@pytest.hookimpl(trylast=True)
+def pytest_sessionstart(session):
+    session._setupstate = RerunAwareSetupState()
 
 
 def _get_resultlog(config):
@@ -206,6 +227,18 @@ def _remove_failed_setup_state_from_session(item):
     setup_state.stack = list()
 
 
+def _should_cause_rerun(item, report):
+    is_terminal_error = _should_hard_fail_on_error(item.session.config, report)
+    xfail = hasattr(report, "wasxfail")
+    should_not_cause_rerun = (
+        item.execution_count > item.allowed_reruns
+        or not report.failed
+        or xfail
+        or is_terminal_error
+    )
+    return not should_not_cause_rerun
+
+
 def _should_hard_fail_on_error(session_config, report):
     if report.outcome != "failed":
         return False
@@ -242,20 +275,14 @@ def pytest_runtest_protocol(item, nextitem):
 
     need_to_run = True
     while need_to_run:
+        item.will_be_rerun = False
         item.execution_count += 1
         item.ihook.pytest_runtest_logstart(nodeid=item.nodeid, location=item.location)
         reports = runtestprotocol(item, nextitem=nextitem, log=False)
 
         for report in reports:  # 3 reports: setup, call, teardown
-            is_terminal_error = _should_hard_fail_on_error(item.session.config, report)
             report.rerun = item.execution_count - 1
-            xfail = hasattr(report, "wasxfail")
-            if (
-                item.execution_count > reruns
-                or not report.failed
-                or xfail
-                or is_terminal_error
-            ):
+            if not _should_cause_rerun(item, report):
                 # last run or no failure detected, log normally
                 item.ihook.pytest_runtest_logreport(report=report)
             else:
@@ -278,6 +305,12 @@ def pytest_runtest_protocol(item, nextitem):
         item.ihook.pytest_runtest_logfinish(nodeid=item.nodeid, location=item.location)
 
     return True
+
+
+def pytest_runtest_makereport(item, call):
+    report = TestReport.from_item_and_call(item, call)
+    item.will_be_rerun |= _should_cause_rerun(item, report)
+    return report
 
 
 def pytest_report_teststatus(report):